requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
